move train_bc to separate module
