+++ act/training.py
@@ -1,0 +1,117 @@
+import torch
+import numpy as np
+import os
+from copy import deepcopy
+from itertools import repeat
+from tqdm import tqdm
+import wandb
+from torchvision import transforms
+
+from general_utils.utils import compute_dict_mean, set_seed # helper functions
+from general_utils.config_handler import make_policy
+from act.policy import make_optimizer
+
+
+def forward_pass(data, policy):
+    image_data, qpos_data, action_data, is_pad = data
+    image_data, qpos_data, action_data, is_pad = image_data.cuda(), qpos_data.cuda(), action_data.cuda(), is_pad.cuda()
+    return policy(qpos_data, image_data, action_data, is_pad) # TODO remove None
+
+
+def repeater(data_loader):
+    epoch = 0
+    for loader in repeat(data_loader):
+        for data in loader:
+            yield data
+        print(f'Epoch {epoch} done')
+        epoch += 1
+
+
+def train_bc(train_dataloader, val_dataloader, config, eval_bc: callable):
+    num_steps = config['num_steps']
+    ckpt_dir = config['ckpt_dir']
+    seed = config['seed']
+    policy_class = config['policy_class']
+    policy_config = config['policy_config']
+    eval_every = config['eval_every']
+    validate_every = config['validate_every']
+    save_every = config['save_every']
+
+    set_seed(seed)
+
+    policy = make_policy(policy_class, policy_config)
+    if config['load_pretrain']:
+        loading_status = policy.deserialize(torch.load(os.path.join('/home/zfu/interbotix_ws/src/act/ckpts/pretrain_all', 'policy_step_50000_seed_0.ckpt')))
+        print(f'loaded! {loading_status}')
+    if config['resume_ckpt_path'] is not None:
+        loading_status = policy.deserialize(torch.load(config['resume_ckpt_path']))
+        print(f'Resume policy from: {config["resume_ckpt_path"]}, Status: {loading_status}')
+    policy.cuda()
+    optimizer = make_optimizer(policy_class, policy)
+
+    min_val_loss = np.inf
+    best_ckpt_info = None
+    
+    train_dataloader = repeater(train_dataloader)
+    for step in tqdm(range(num_steps+1)):
+        # validation
+        if step % validate_every == 0:
+            print('validating')
+
+            with torch.inference_mode():
+                policy.eval()
+                validation_dicts = []
+                for batch_idx, data in enumerate(val_dataloader):
+                    forward_dict = forward_pass(data, policy)
+                    validation_dicts.append(forward_dict)
+                    if batch_idx > 50:
+                        break
+
+                validation_summary = compute_dict_mean(validation_dicts)
+
+                epoch_val_loss = validation_summary['loss']
+                if epoch_val_loss < min_val_loss:
+                    min_val_loss = epoch_val_loss
+                    best_ckpt_info = (step, min_val_loss, deepcopy(policy.serialize()))
+            for k in list(validation_summary.keys()):
+                validation_summary[f'val_{k}'] = validation_summary.pop(k)            
+            wandb.log(validation_summary, step=step)
+            print(f'Val loss:   {epoch_val_loss:.5f}')
+            summary_string = ''
+            for k, v in validation_summary.items():
+                summary_string += f'{k}: {v.item():.3f} '
+            print(summary_string)
+                
+        # evaluation
+        if (step > 0) and (step % eval_every == 0):
+            # first save then eval
+            ckpt_name = f'policy_step_{step}_seed_{seed}.ckpt'
+            ckpt_path = os.path.join(ckpt_dir, ckpt_name)
+            torch.save(policy.serialize(), ckpt_path)
+            success, _ = eval_bc(config, ckpt_name, save_episode=True, num_rollouts=10)
+            wandb.log({'success': success}, step=step)
+
+        # training
+        policy.train()
+        optimizer.zero_grad()
+        data = next(train_dataloader)
+        forward_dict = forward_pass(data, policy)
+        # backward
+        loss = forward_dict['loss']
+        loss.backward()
+        optimizer.step()
+        wandb.log(forward_dict, step=step) # not great, make training 1-2% slower
+
+        if step % save_every == 0:
+            ckpt_path = os.path.join(ckpt_dir, f'policy_step_{step}_seed_{seed}.ckpt')
+            torch.save(policy.serialize(), ckpt_path)
+
+    ckpt_path = os.path.join(ckpt_dir, f'policy_last.ckpt')
+    torch.save(policy.serialize(), ckpt_path)
+
+    best_step, min_val_loss, best_state_dict = best_ckpt_info
+    ckpt_path = os.path.join(ckpt_dir, f'policy_step_{best_step}_seed_{seed}.ckpt')
+    torch.save(best_state_dict, ckpt_path)
+    print(f'Training finished:\nSeed {seed}, val loss {min_val_loss:.6f} at step {best_step}')
+
+    return best_ckpt_info
